[back][Dev][cache implemented and used for getUser]
